Make shell responsible for activity ID generation
Enables tracking of activity launch throughout the whole process, so that
the shell can be aware of the activity ID from the moment the activity
is started by the shell, until the activity becomes active.  Previously,
the activity itself generated its own ID and told the shell what it was.