Expose owner details through the Shell's DBus service
For security, we need the PresenceService to listen for changes to the
owner's attributes, like changed color, nickname, icon, and current activity,
rather than having D-Bus API in the PS itself that any process could call.

So, the shell provides signals when these attributes change, which the PS
listens to and pushes out over the network accordingly.