Make debugging tp-gabble easier
When the environment variable TP_DEBUG is set, the sugar shell
will print out the session bus address it's using, then poll that bus
every 5 seconds waiting for a gabble instance to appear.  This allows
you to launch gabble in another window for debugging with LM_DEBUG and
gdb, which wasn't possible before because the sugar emulator uses
a custom session bus.