Support moving of files to datastore when using write_file()
Using the transfer_ownership argument, activities using the default activity
datastore integration methods (namely write_file) will now tell the datastore
that it can move the files by default.  This reduces the copies required, which
is slow on flash.  For activities not using the standard APIs (Record, etc),
the datastore bindings allow the activity to specify when ownership should
transfer.