Reset activity icons when clicked/launched (#7273, #7274)
The act of clicking on an activity icon in Home to launch it
switches the view to the launcher, eliminating the expected
hovering-changed event (leave).  This commit hides the palette
instantly (it used to remain visible in the launcher) and
un-colors the icon (which used to remain colored when returning
to Home).