#7248 speaker icon fullness more intuitive
The speaker icon's "fullness" is now more intuitive in the same way as
the battery icon's in #4208: the icon fullness is chosen by finding
the closest icon starting from the current level going *down* to zero.
For example: rather than, given four icons of -000, -033, -066, and
-100 percent "fullness", the -100 icon being chosen if the speaker
volume level was 67, the -066 icon would be chosen.  This
"underestimates" the volume, which has been identified as more
intuitive.

|TestCase|

Set the volume to less than 100%.  The icon should not be full.