Filesystem paths should be byte arrays, not unicode sequences. (dlo trac #7733)
Passing a unicode string to os.path.join(x,y) and other functions will attempt
to convert all sides of the path to unicode using the ASCII encoding.  This
will prohibit the use of non-ASCII characters in filenames, which is not
what we want!  We need all paths passed around internally to be 'str' objects,
not 'unicode' objects.  Unfortunately, dbus' STRING type enforces UTF-8
encoding.  So, we compromise: we use 's' in our dbus signature, but we pass the
'utf8_strings=True' keyword option when we declare our methods, so that
the parameters are given to use as 'str' not 'unicode'.

These functions will not support filenames which are not UTF-8 encoded.