Avoid network_appeared race
This prevents access points being shown twice (#7415). See the comments
in the patch for explanation.

It is perhaps questioanable why NetworkManager is producing
NetworkAppeared messages at this time, but it doesn't seem worth the
effort digging into that side of things given that we will soon be
upgrading to NetworkManager-0.7 which will totally shake up this code
anyway.