#7685 Redo the "wordless" palette to use RadioToolButton instead of a menu.
Eben wants to use the menu part in the future to switch between "new activity"
and "resume document" modes in the favorites view, so he prefers we use
the 'custom content' section of the Palette.  This also would let us switch
to a multi-row tabular layout in the future as well, although we don't have
enough layout options at the moment to make it worthwhile.