Trac #7992: don't waste power blinking the cursor.
Original reporter: "[Sugar control panel] includes a search box with a
blinking cursor, which makes the process wake up every second to blink
it. ... This burns power unnecessarily."

I've used the gtk-cursor-blink-timeout option, so that we stop
blinking after 3s. This preserves the UI indication that interaction
is possible. The full set of possible settings is at
http://library.gnome.org/devel/gtk/unstable/GtkSettings.html#GtkSettings--gtk-cursor-blink-timeout
if we decide to tweak this (slower blink, blink longer before stopping,
don't blink at all, etc).

This doesn't seem to affect the text box widgets in Browse, but does affect
most (all?) GTK applications/activities.