CanvasPulsingIcon: Don't begin pulse loop on resume if not pulsing
When switching to the Neighborhood view, the system starts waking up 10
times per second *per AP*, even though nothing on the screen is being
updated.

This is because meshbox attemps to unpause all its AP pulsing icons,
and CanvasPulsingIcon makes the incorrect assumption that an icon being
resumed was previously pulsing.

Add a check that only resumes the pulse loop if the icon was already
pulsing, and also (for safety and clarity) prevent the pulse loop from
starting on a paused icon.