View Source: Make document the default source
Toolbar:
document available: [document, bundle, title]
no document available: [bundle, title]

Icons:
The coloured activity icon represents the source of the activity instance. The
outlined activity icon represent the bundle source for the activity itself,
the template. We use this to help draw a distinction between the
"activity instance" and the "activity bundle"