Activate windows when changing to activity view
This patch fixes a bug where pressing F4 immediately after starting
Sugar does not switch to the Journal.

This happened because the Journal has not yet been activated, so when
the desktop is hidden in response to pressing F4, there is no active
window to take over the display. The journal starts in iconified state.

We also have to be careful about propogating correct X event times
for the F4 keypress case (this is not an event that GTK+ handles,
it comes from KeyGrabber).