Journal entry bundles shouldn't create a 2nd ds entry
This fixes a problem where "Show in Journal" from a Browse download
was broken. This is because the JEB installation code would extract the
content from the bundle entry, install it as a new datastore entry,
and then delete the bundle entry.

This is fixed by making the journal entry from the bundle take over
the datastore entry that belonged to the bundle.