Configure fonts through GConf (#1584)
Sugar currently configures fonts through gtkrc and through the
sugar.graphics.style class, both of which are hard to modify by
implementors/deployers.

Move this into GConf and apply these settings to the sugar shell
at startup.

In the longer term, Sayamindu's efforts to create a settings manager
will build on this work.