Revert "Move to cjson and drop pyjson and simplejson"
This reverts commit ee4535c98ae74347e7072909d49dcf8a5e16ca7b.

cjson has a big bug dealing with slashes, this is a significant
long-term bug and upstream has not been responsive other than
acknowledging it. This bug breaks journal entry bundles.
http://dev.sugarlabs.org/ticket/1553

Thanks to Martin Langhoff for identifying and researching this issue

also merged in 738efb8179c9f52df5cf2e21de0acda99d09dccf

Switch entirely to simplejson

I misunderstood the state of affairs with the numerous json parsers.
Whatever we used to have as 'json' isn't really available on Py2.6.

Move to simplejson which seems to be the most functional implementation,
available on both old and new setups.