Fix duplication of OLPC mesh icons
Three or more mesh icons were being added after every suspend/resume.

There was already code present to detect when the device goes away and
remove the icons, but it was not working. NetworkManagerObserver was
attempting to look at the properties of the gone-away device in order to
see if it is a mesh device, but this was failing because the device has
already been destroyed at this point.

Solve this by recording the device path of the mesh device when it is
detected, so that we can later determine if the gone-away device is the
mesh device.

Signed-off-by: Martin Abente <mabente@paraguayeduca.org>