add restart option configurable via gconf (#1206)
Nepal requested reordering and removal of restart, which was patched for
them in 0.82 and implemented in 0.84.

Paraguay and Uruguay detected the change in 0.84 and requested the
restart option be restored.

This patch restores the restart option as a configurable option, using
the gconf key /desktop/sugar/show_restart, default true.

Refer to http://bugs.sugarlabs.org/ticket/1206 for discussion.

Acked-by: Tomeu Vizoso <tomeu.vizoso@collabora.co.uk>