Sugar Ad-hoc icons show in search results when connect/disconnect to AP OLPC #10412
When searching for an AP the Sugar ad-hoc networks are greyed
out (if their name does not contain the search string). When
then trying to connect to an AP the Sugar ad-hoc networks get
colored again.

The patch does make sure that the color is not set by each update
and that the member change does not trigger setting the color
of the icon.

Tested-By: Sascha Silbe <sascha-pgp@silbe.org>