Possibility to share Journal entry using a storage device #9657
The metadata and the preview for each entry are stored
in a hidden folder named .Sugar_metadata on the device.

The data files will be stored in the root folder of
the device. Like in 0.82 the entries will be colored
in the creator's color.

Note: Entries that do not contain a data file can not
be copied to a storage device anymore in 0.84
(this was possible in 0.82), an alert is displayed
to the learner in those cases. This change has been
independent of these patches.

The code is written defensive. If the metadata/preview
is corrupted or has been removed it will be skipped,
the deletion of a data file on the device is
handled gracefully.