Convert Journal entries that have been saved to a storage device in 0.82 #9658
This code will convert the old Journal entries to the new
scheme described above. When a device containing a folder .olpc.store
is plugged into the machine the conversion will happen
automatically in the background. Files that have the metadata converted
already from a previous run will be omitted from the conversion.
The .olpc.store won't be removed after completion to be able
to restore if anything goes wrong during conversion.

Note: Entries that do not contain a data file can not
be copied to a storage device anymore in 0.84
(this was possible in 0.82), for code simplification
we do not restore those entries from 0.82.

The conversion code is written defensive. If an entry can
not be converted it will be skipped, if the database is
broken the conversion is skipped etc. So the deletion of a data
file on the device or any corruption to the .olpc.store
should not cause a crash to Sugar.