Copy a Journal entry including the metadata to a removable device
The new code will store the metadata and preview in a hidden
folder like in 0.82. The folder will be named .Sugar_metadata.
There will be two files per entry, one for the metadata named
filename.metadata and one for the preview (if exists)
named filename.preview. The data files will be stored in the
root folder of the device. User experience: When the user views
the content of a mass storage device Journal entries that have
been copied to the device will be shown in the color of
the creator (same behavior as with 0.82).

Note: In 0.84 entries that do not contain a data file can
not be copied to a storage device anymore
(this was possible in 0.82), an alert is displayed to the
learner in those cases. This change has been independent
of these patches.

The code is written defensive. If the metadata/preview is
corrupted or has been removed it will be skipped, so the
deletion of a data file on the device is handled gracefully.

The learner has now the options to: backup the Journal
entry to a storage device and copy it back to his own Journal
later, share his entry with another learner over a storage
device and use the file on a non Sugar machine
(showing a picture to your friend on his computer).

more info: http://wiki.sugarlabs.org/go/Features/Journal_Entry_Sharing

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Sascha Silbe <silbe@activitycentral.com>