Sugar Ad-hoc network: reconnect after suspend SL #2187
Setting the 'autoconnect' property for the connection
will let NM do the reconnection when the device is up. If
the Sugar Ad-hoc network is around it will automatically
connect to it. The connection is not stored after logging
out of Sugar.

The autoconnect functionality inside Sugar is still needed
for the cases when we log into Sugar, after a timeout we
will connect to an exisiting Ad-hoc network or will default
to the Sugar Ad-hoc network on channel 1. I increased the
timeout to 60 seconds because during testing I found to
get more reliable results like that.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-By: Sascha Silbe <silbe@activitycentral.com>