Cache the XO palette in the Home View, part of #2726
Since we do not cache the palette for the BuddyIcon anymore,
we as well do not cache the Owner palette in the Favorites view.
The current code tries to change the register menu (school server)
in the palette when successfully registered (label it 'register again').
But the palette has been destroyed and therefore this attempt
fails.

The patch does cache the palette of the Owner, this has been
the behavior before 356641c332d6cc889b265dfc29598004cf37763c. As
registering (and therefore the change of the palette) happens
rather seldom compared to the change of a buddy palette, I think
the Owner palette can be considered to be mostly static
and therefore we can cache it.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Sascha Silbe <silbe@activitycentral.com>