Fix inviting from friends view OLPC #10767
Add the contact_id and the account information to the buddymodel used in
the group view. We do not store it in the friends model as the contact_id and
the account information may vary. When we read the friends information from
disk the contact_id and account information gets added to the model whenever
we have it.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-By: Sascha Silbe <silbe@activitycentral.com>