Make sure the entry has the right color when copying from external device to Journal SL #2690
The patch does make sure that the color of the entry has the owner's color when you copy
it from an external device to the Journal. It handles the case where when you copy a Journal
entry from another user on an external device and then into your Journal cleanly.
The color stays the same.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-By: Sascha Silbe <silbe@activitycentral.com>