Friendstray: make sure the tray is right on the sharer side OLPC #10817
The code path that listens for the 'activity-added' signal is used to
track the following use case: 'A' starts an activity
(he gets the 'active-activity-changed' signal from the shell), he then
at a later point shares the activity. In order to track joining/leaving
buddies he needs to get the shared activity. On other machines if we
receive the 'activity-added' signal we can ignore it, therefore the
check if the current active activity is the shared one.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-By: Sascha Silbe <silbe@activitycentral.com>