Prevent view source display of some binary and "junk" files (SL#2854)
This patch addresses the issue raised in SL#2854: the exclusion of some file
types from being displayed by viewsource. This version of the patch uses a
global, _EXCLUDE, that includes a list of many more common file types to
exclude, as per some mining of .gitignore files by Sascha. In a future patch,
it may make sense to append the data from the .gitignore and git exclude files.

[added exclude list for (exact-match) file names, tweaked description]
Signed-off-by: Sascha Silbe <silbe@activitycentral.com>
Acked-by: Simon Schampijer <simon@laptop.org>