Group members of a shared activity around it OLPC #10675
Like in previous versions we use the shell model and listen if the
current-activity changes. We do then call 'SetCurrentActivity' and
let the other buddies know. If we started the activity and share it
later we listen on the 'ActivitiesChanged' signal and do call
'SetCurrentActivity' then so the other buddies can update their views
accordingly.

This patch also removes the dependency in the shell on the
neighborhood model which was a hack to determine the color of
a shared activity. The neighborhood view now updates the color
accordingly and that information is transferred to the shell.

The patch furthermore removes the tracking of the own buddy when
we receive 'ActivitiesUpdated' in the neighborhood model. With the
changes from above this triggered an error like described in
OLPC #10671. A few checks have been added to make sure that #10671
does not occure anymore.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Sascha Silbe <silbe@activitycentral.com>