Network device icon: use "Disconnect" label (SL#3007)
The current label "Disconnect..." has an ellipsis for no apparent
reason. The equivalent menu on the neighborhood view has "Disconnect".
Remove the ellipsis to make these consistent.

This will also fix a bad Spanish translation of "Disconnect..."
("Desconectando..." which means "Disconnecting..."), as the Spanish
translation for Disconnect ("Desconectar") is correct.