Add duplicate functionality to the Journal and enhance copy functionality
This patch adds a duplicate option to the Journal entry palette and
the entry detail view. This will replace the keep button functionality
from the activity toolbar. The keep button will be deprecated in another
patch for the toolkit.

The copy option which copied previously to the clipboard by default
has been enhanced to allow: copying to the clipboard, copying to an
external device, copying from an external device to the Journal and
copying between external devices. Copying to the clipboard is now
a visible option in the menu. In the detail view the palette pops up
when doing a left click and showing the available options instead of
copying directly to the clipboard.

The design discussion has been taking place
at: http://lists.sugarlabs.org/archive/sugar-devel/2011-May/031316.html

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-By: Sascha Silbe <silbe@activitycentral.com>