Do startup animation of the activity icon using scaling and alpha - v5
This render the SVG image only one time, and then use alpha and scaling
of the rendered icon and will speed up the overall activity startup.
Another side effect is that the startup zoom not visible in many
activities, is visible now - #2080
The MIN_ALPHA value was changed to do the effect more visible.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>