Add support for XO 1.75
The hardware specific information has been moved from '/ofw' to
'/proc/device-tree'. This has an effect on the Control Panel
where we display the serial number and the firmware number.
Furthermore the registration code needs to be aware of the possible
new place in order to send the serial number and UUID during
authentication.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Reviewed-by: Gonzalo Odiard <gonzalo@laptop.org>
Tested-by: Gonzalo Odiard <gonzalo@laptop.org>