adding $HOME/Documents to volumestoolbar
This patch adds $HOME/Documents to the volume toolbar in the Journal view.
The rationale is to make it easier for people to move files in and out of
the data store from the file system -- a feature oft requested by teachers.
It also means that Sugar activities can more readily access files generated
outside of Sugar -- another feature requested by teachers.

This resubmission of the patch includes a suggestion by Silbe that the
xdg-user-dir call be factored out in a separate (global) function

Note that this patch requires the inclusion of a new icon, user-documents,
that is included in a separate patch.

Signed-off-by: Walter Bender <walter.bender@gmail.com>
Acked-by: Simon Schampijer <simon@laptop.org>