Add clone to view-source submenu
This patch adds a Clone submenu to the view source toolbar, enabling the
cloning of an activity for end-user modification.

As per the design team meeting, this patch includes the generation of an
xo bundle to be copied to the Journal of the cloned activity.

One known bug not addressed here is that if the bundle_id is missing, the
cloning will be successful, but the resultant activity will not launch.

Signed-off-by: Walter Bender <walter.bender@gmail.com>
Acked-by: Simon Schampijer <simon@laptop.org>