Invitations: remove item from tray after using the join/decline palette options, SL #3089
The old code did only handle the case cleanly when you
clicked on the icon in the tray. When using the palette options
several errors occured.

This patch does add a change in behavior: it reveals the
palette when you click on the button. This is to ensure
that a user is aware of the available options (discussed with
Gary).

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-By: Sascha Silbe <silbe@activitycentral.com>