FileTransfer: handle cancellation cleanly, OLPC #11065
When a file transfer has been canceled (the sender quits
before the receiver transferred it) we show a cancellation
message and the option to dismiss the transfer status
palette.

The behavior for a cancellation by the sender is a bit
different in telepathy as noted in [1]. You do only
get the cancellation message and the option to dismiss
the transfer status palette once you accepted the
transfer. The same behavior is true in Empathy.

The cancellation message is the same as in Empathy.

The resume functionality has never been offered by
telepathy.

[1] https://bugs.freedesktop.org/show_bug.cgi?id=20621

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-By: Sascha Silbe <silbe@activitycentral.com>