SpreadLayout: fix grid cell allocation
The calculation of how many grid cells to assign a child element
has a bug in that it rounds down instead of up.

With a grid cell size of 4x4, if a child element of 9 pixels in size
is added, the current code calculates that it needs 2.25 grid cells,
but then rounds this down to 2. When the child gets drawn it will
then draw outside of the grid cell bounds it was given.

Fix this by rounding up instead of down: in the above example, 3 grid
cells would be allocated. The canvas elements used already centre
themselves within their allocation.

In current Sugar, I can't see any notable differences from this change.
However, this is a fairly obvious fix, and a necessary part of the
later solution to the issue where the owner icon zoom animation does not
quite match the location of the owner icon on the views.

Acked-By: Sascha Silbe <silbe@activitycentral.com>
Signed-off-by: Daniel Drake <dsd@laptop.org>