Control Panel: making about my computer section hardware independent, OLPC #11232
'/ofw' is not used anymore on the XO-1.75, '/proc/device-tree' is used now
instead. Instead of just adjusting for that change we took the
chance to make the section hardware independent.

As the firmware version we display the bios version if available on non XO
hardware. On the XO we display a human readable extraction of what is
found in .../openprom/model. On other architectures (e.g. powerpc) we
display all of the content of that file.

As ethtool has become a depedency of Sugar we can now
display the wireless firmware as well on any hardware.

The serial number is often only readable by root on non-XO
hardware, that is why we do not read it. If the serial number
can not be found we display the entry as 'Not available'.

The patch has been tested on XO 1, 1.5, 1.75 and on a Thinkpad T61.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Sascha Silbe <silbe@activitycentral.com>
Reviewed-by: James Cameron <quozl@laptop.org>