Make sure we use the Sugar cursor in the intro screen, OLPC #10712
As stated in the ticket Metacity sets the cursor after
it's startup, so we need to make sure Metacity is already
running before we set the cursor. This is what the current
code has been doing. We do show the intro screen before
the window manager may be available. For a non invasive
bug fix that avoids races we just call 'xsetroot -cursor_name'
another time before the intro screen is displayed.

We have a long history of setting the Sugar cursor with many
races. To improve the situation SL #3204 is filed as a
follow up.

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Sascha Silbe <silbe@activitycentral.com>