Journal detail view: don't choke on invalid 'keep' property (fixes SL#1591)
Properties of data store entries can get corrupted, e.g. due to low level
crashes or running out of battery (see OLPC#11372 [1] for a real-life
example). In addition any activity can - accidentally or on purpose - write
data store entries with arbitrary metadata.

By comparing the 'keep' property as a string we can avoid the ValueError that
might happen when trying to convert the property value to an integer.

[1] https://dev.laptop.org/ticket/11372

Reported-by: Gary Martin <garycmartin@googlemail.com>
Signed-off-by: Sascha Silbe <silbe@activitycentral.com>
Acked-by: Simon Schampijer <simon@laptop.org>