Implementation of Text to Speech Feature [1]
The patch adds a device icon to the frame with controls
to play/pause/stop any text that is selected by the user in
the Sugar UI. The palette does also have options to adjust
the pitch and rate of the spoken text. These settings are
stored in the GConf backend, the two keys are added to the
Sugar schema.

As backend gstreamer-plugins-espeak is used [2], it has been
in use in many activities already and is packaged for Fedora,
it will be added as a dependency to Sugar.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>

[1] http://wiki.sugarlabs.org/go/Features/GlobalTextToSpeech
[2] http://wiki.sugarlabs.org/go/Activity_Team/gst-plugins-espeak