Prevent handling network devices twice (OLPC#11782)
Due to the asynchronous way that dbus works, it is possible to
receive a network device listed in the initial GetDevices call,
and as a DeviceAdded signal, if the device appears during Sugar
initialisation.

Check to see that a device is really new before processing it.
Fixes a case seen by Manuel and me where two network device
icons appear.

Signed-off-by: Daniel Drake <dsd@laptop.org>