Only add one time every object to the clipboard, SL #3371
This patch changes the behaviour of the clipboard tray,
every object is only added one time, if it already exists,
the already added object is selected and a notification is displayed
(showing a flashing icon in the lower left corner).

A hash over the data in the object is used to determine if the object
is already present in the Journal. When the object is an uri,
a md5sum is calculated over the file and the hash is done over the result.

We need check if the object is already in the clipboard
because a bad interaction between the clipboard in write and
the text to speech feature. See the ticket for more information.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>