Release audio device after text to speech - OLPC #11829
Since on the XO we do not have a sound serever (e.g. pulse audio)
we need to make sure to release the audio device in order to make
it usable by other components like activities.

Gstreamer does wait for 'gst.MESSAGE_EOS' message in order to release
the device. See the documentation at [1][2].
[1] http://gstreamer.freedesktop.org/data/doc/gstreamer/head/manual/html/chapter-helloworld.html#section-helloworld
[2] http://pygstdocs.berlios.de/pygst-tutorial/playbin.html

Signed-off-by: Simon Schampijer <simon@laptop.org>
Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>